refactor useSelectedTermsOfSavedPost hook to ts
